requirements.txt
runpod==1.7.7
python-dotenv==1.0.1
numpy==1.26.4
Pillow==10.1.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
transformers==4.44.2
sentencepiece==0.1.99
https://github.com/Dao-AILab/flash-attention/releases/download/v2.6.2/flash_attn-2.6.2+cu123torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl
decord
librosa==0.9.0
soundfile==0.12.1
vector-quantize-pytorch==1.18.5
vocos==0.1.0
moviepy
gradio